Add method DataRegionSelection.getSelectedIntegers() and eliminate lots of duplicate String -> Integer parsing code
TextWriter now implements AutoCloseable -> switch callers to use try-with-resources

Switch to try-with-resources elsewhere, eliminate unused imports and other warnings

+++ microarray/src/org/labkey/microarray/controllers/FeatureAnnotationSetController.java
@@ -365,7 +365,7 @@ public class FeatureAnnotationSetController extends SpringActionController
         {
             if (_singleObjectRowId != null)
             {
-                return new int[] { _singleObjectRowId.intValue() };
+                return new int[] {_singleObjectRowId};
             }
             return PageFlowUtil.toInts(DataRegionSelection.getSelected(getViewContext(), clear));
         }
